\name{capitalize}
\alias{capitalize}
\title{Capitalize the First Letter of Each Word}
\description{
  This function capitalizes the first letter of each word in a string.
}
\usage{
capitalize(x)
}
\arguments{
  \item{x}{A character vector of strings to be capitalized.}
}
\value{
  A character vector with the first letter of each word capitalized.
}
\examples{
capitalize("this is a test")
}